Complexity Measures

complexity computes Diversity/Ubiquity and Country/Product Complexity Indices

complexity(rca, country = "country", product = "product",
  value = "value", method = "fitness", iterations = 20,
  extremality = 1, tbl = TRUE, atlas = FALSE)

Arguments

rca
matrix or data.frame with RCA values

country
column containing countries (applies only if d is a data.frame)

product
column containing products (applies only if d is a data.frame)

value
column containing traded values (applies only if d is a data.frame)

method
which method to use (by default is "fitness", it can also be "reflections" or "eigenvalues")

iterations
number of iterations to use in the fitness, reflections and indirectly in eigenvalues method (by default is 20)

extremality
numeric coefficient to use in the fitness method (by default is 1)

tbl
TRUE (default) returns a data.frame and FALSE returns a matrix

atlas
remove the countries not ranked in The Atlas of Economic Complexity (by default is FALSE)

Value

A list with four data frames or matrices.

Details

Given a \(C\times P\) matrix (C for "countries" and P for "products") or an equivalent 3-columns data frame with RCA values as input, this function implements the equations: $$\text{(Diversity)}\: k_{c}^{(0)} = \sum_p R_{cp}$$ $$\text{(Ubiquity)}\: k_{p}^{(0)} = \sum_c R_{cp}$$ $$\text{(Fitness)}\: \tilde{F}^{(n)}(\gamma) = \sum_{p} R_{cp}Q_{p}^{n-1}$$ $$\text{(Complexity)}\: \tilde{Q}^{(n)}(\gamma) = \left[\sum_{c} R_{cp}(F_{c}^{n-1})^{-\gamma}\right]^{-1/\gamma}.$$

Besides this implementation, the function offers the reflections method and eigenvalues computation as alternatives to the Fitness-Complexity formulation.

References

For more information on complexity measures see:

Mariani M, Vidmer A, Medo M, Zhang Y (2015). “Measuring economic complexity of countries and products: which metric to use?” The European Physical Journal B, 88(11), 293. doi: 10.1140/epjb/e2015-60298-7 .

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

and the references therein.

See also

Examples

complexity(ec_output_demo$rca_tbl)

#> $country_complexity
#> # A tibble: 158 x 2
#>    country value
#>    <chr>   <dbl>
#>  1 deu     28.8 
#>  2 fra     21.4 
#>  3 usa     14.1 
#>  4 nld     11.5 
#>  5 che     10.5 
#>  6 gbr      9.81
#>  7 ita      9.61
#>  8 swe      7.35
#>  9 aut      7.34
#> 10 jpn      7.27
#> # ... with 148 more rows
#> 
#> $product_complexity
#> # A tibble: 991 x 2
#>    product value
#>    <chr>   <dbl>
#>  1 7111     75.1
#>  2 2850     55.9
#>  3 8804     32.1
#>  4 9307     24.7
#>  5 9112     20.2
#>  6 3505     19.6
#>  7 3813     14.3
#>  8 7507     12.6
#>  9 2927     12.1
#> 10 2810     11.8
#> # ... with 981 more rows
#> 
#> $diversity
#> # A tibble: 158 x 2
#>    country value
#>    <chr>   <dbl>
#>  1 afg        28
#>  2 ago        52
#>  3 alb        32
#>  4 and        23
#>  5 ant        23
#>  6 are         1
#>  7 arg        83
#>  8 aus        81
#>  9 aut       290
#> 10 bdi        16
#> # ... with 148 more rows
#> 
#> $ubiquity
#> # A tibble: 991 x 2
#>    product value
#>    <chr>   <dbl>
#>  1 0101       19
#>  2 0102       26
#>  3 0103       16
#>  4 0104       19
#>  5 0105       11
#>  6 0106       35
#>  7 0201       21
#>  8 0203       15
#>  9 0204        6
#> 10 0205       12
#> # ... with 981 more rows
#>