Complexity Measures

complexity_measures() computes the Economic Complexity Index and the Product Complexity Index.

Usage

complexity_measures(
  balassa_index,
  method = "fitness",
  iterations = 20,
  extremality = 1
)

Arguments

balassa_index

(Type: matrix) the output from balassa_index()) or an equivalent arrangement.

method

(Type: character) one of these methods: fitness, reflections or eigenvalues. By default this is set to "fitness".

iterations

(Type: numeric) the number of iterations to use. By default this is set to 20.

extremality

(Type: numeric) the parameter to use in the fitness method. The other methods don't use this parameter. By default this is set to 1.

Value

A list of two named numeric vectors.

Details

The current implementation follows (Mariani et al. 2015) to obtain different alternatives that account for diversification in bipartite relations.

References

For more information on this index see:

Mariani M, Vidmer A, Medo M, Zhang Y (2015). “Measuring economic complexity of countries and products: which metric to use?” The European Physical Journal B, 88(11), 293. doi:10.1140/epjb/e2015-60298-7 .

and the references therein.

Examples

co <- complexity_measures(economiccomplexity_output$balassa_index)

# partial view of indexes
n <- seq_len(5)
co$complexity_index_country[n]
#>        afg        ago        aia        alb        and 
#> 0.78605655 0.03999516 1.05645538 1.24261128 1.37321261 
co$complexity_index_product[n]
#>      0011      0012      0013      0014      0015 
#> 0.7538883 0.7491391 2.3018690 0.9903918 1.3120378