Complexity Outlook

complexity_outlook() computes the Complexity Outlook Index and the Complexity Outlook Gain.

complexity_outlook(balassa_index, proximity_product, complexity_index_product)

Arguments

balassa_index
(Type: dgCMatrix) the output from balassa_index()) or an equivalent arrangement.

proximity_product
(Type: dgCMatrix) the output from proximity()) or an equivalent arrangement.

complexity_index_product
(Type: numeric) the output from complexity_measures()) or an equivalent arrangement.

Value

A list of two named numeric vectors.

Details

The current implementation follows (Hausmann et al. 2014) to measure unexploited export oppportunities.

References

For more information on this index see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

and the references therein.

Examples

complexity_outlook(
 economiccomplexity_output$balassa_index,
 economiccomplexity_output$proximity$proximity_product,
 economiccomplexity_output$complexity_measures$complexity_index_product
)

#> $complexity_outlook_index
#>                     Bepis 9                  Bird World 
#>                1.7444398507                0.0405825586 
#>                  Crepktct 4                 Earth C-137 
#>                1.5696522272                0.0851599657 
#> Galactic Federation Outpost  Galactic Federation Prison 
#>                0.8536235071                1.5087409922 
#>                  Gear World                Glaagablaaga 
#>                2.8023317343                0.9811072446 
#>              Gromflom Prime 
#>                0.0008766251 
#> 
#> $complexity_outlook_gain
#> 9 x 12 Matrix of class "dgeMatrix"
#>                                alpha         beta    delta   epsilon       eta
#> Bepis 9                     2.927624 0.0010937760 2.927624 0.0000000 0.1131705
#> Bird World                  2.928597 0.0010864115 2.928597 0.7504575 0.1141356
#> Crepktct 4                  2.926295 0.0010864115 2.926295 0.7489167 0.1136155
#> Earth C-137                 0.000000 0.0002245155 0.000000 0.0000000 0.1009748
#> Galactic Federation Outpost 2.928350 0.0010864115 2.928350 0.0000000 0.1133684
#> Galactic Federation Prison  2.927727 0.0000000000 2.927727 0.7500228 0.1141356
#> Gear World                  2.921526 0.0010937760 2.921526 0.0000000 0.0000000
#> Glaagablaaga                2.922499 0.0010864115 2.922499 0.7337949 0.0000000
#> Gromflom Prime              2.928597 0.0000000000 2.928597 0.7504648 0.1141430
#>                                  gamma      iota      kappa   lambda       mu
#> Bepis 9                     0.04042168 0.1647654 0.06667830 1.471275 1.060685
#> Bird World                  0.00000000 0.1647581 0.06690333 1.472240 1.062623
#> Crepktct 4                  0.00000000 0.1642380 0.00000000 0.000000 1.060568
#> Earth C-137                 0.03549475 0.1586676 0.05904625 0.000000 0.000000
#> Galactic Federation Outpost 0.00000000 0.1642380 0.00000000 1.471473 1.061608
#> Galactic Federation Prison  0.00000000 0.1647581 0.06690333 1.471371 1.062043
#> Gear World                  0.03106237 0.1444161 0.05107945 1.465177 0.000000
#> Glaagablaaga                0.00000000 0.0000000 0.00000000 1.465622 0.000000
#> Gromflom Prime              0.04057595 0.1647654 0.06692542 1.472248 1.062630
#>                                 theta     zeta
#> Bepis 9                     0.0000000 1.469432
#> Bird World                  0.9995132 1.470397
#> Crepktct 4                  0.9979790 1.468095
#> Earth C-137                 0.0000000 0.000000
#> Galactic Federation Outpost 0.9987718 1.470150
#> Galactic Federation Prison  0.9989337 0.000000
#> Gear World                  0.0000000 1.463334
#> Glaagablaaga                0.9778165 1.464299
#> Gromflom Prime              0.9995132 1.470295
#>