Countries position

Countries position

countries_position(rca = NULL, country1 = "country",
  product1 = "product", value1 = "value", product_proximity = NULL,
  product21 = "from", product22 = "to", value2 = "value",
  product_complexity_index = NULL, product3 = "product",
  value3 = "value", tbl_output = FALSE)

Arguments

rca
matrix or tibble/data.frame (e.g. the output of revealed_comparative_advantage()). If the input is a matrix it must be a zero/one matrix with countries in rows and products in columns. If the input is a tibble/data.frame it must contain at least three columns with countries, products and values.

country1
string to indicate the column that contains exporting countries in rca (set to "country" by default)

product1
string to indicate the column that contains exported products in rca (set to "product" by default)

value1
string to indicate the column that contains traded values in rca (set to "value" by default)

product_proximity
matrix or tibble/data.frame (e.g. the output of proximity_matrices()). If the input is a matrix it must be a numeric matrix with products in both rows and columns. If the input is a tibble/data.frame it must contain at least three columns columns with products (twice) and values.

product21
string to indicate the first column that contains exported products in product_proximity (set to "from" by default)

product22
string to indicate the second column that contains exported products in product_proximity (set to "to" by default)

value2
string to indicate the column that contains proximity values in product_proximity (set to "value" by default)

product_complexity_index
numeric vector or tibble/data.frame complexity_measures()). If the input is a vector it must be numeric with optional names. If the input is a tibble/data.frame it must contain at least two columns with products and values.

product3
string to indicate the column that contains exported products in product_proximity (e.g. "product")

value3
string to indicate the column that contains proximity values in product_proximity (e.g. "value") from complexity_measures())

tbl_output
logical value to use tibble output instead of a matrix output (set to FALSE by default)

References

For more information on proximity distance, complexity outlook, complexity outlook gain and its applications see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

Examples

countries_position(
  rca = package_output_demo$revealed_comparative_advantage_matrix,
  country1 = "country",
  product1 = "product",
  value1 = "value",
  product_proximity =
   package_output_demo$proximity_matrix$product_proximity,
  product21 = "from",
  product22 = "to",
  value2 = "value",
  product_complexity_index =
   package_output_demo$complexity_measures_numeric$product_complexity_index,
  product3 = "product",
  value3 = "value",
  tbl_output = TRUE
)

#> $proximity_distance
#> # A tibble: 880 x 3
#>    country    product          value
#>    <chr>      <chr>            <dbl>
#>  1 Albania    1 Transportation 0.674
#>  2 Algeria    1 Transportation 0.395
#>  3 Angola     1 Transportation 0.893
#>  4 Argentina  1 Transportation 0.449
#>  5 Armenia    1 Transportation 0.392
#>  6 Aruba      1 Transportation 0.893
#>  7 Australia  1 Transportation 0.674
#>  8 Azerbaijan 1 Transportation 0.639
#>  9 Bahamas    1 Transportation 0.893
#> 10 Bangladesh 1 Transportation 0.580
#> # … with 870 more rows
#> 
#> $complexity_outlook
#> # A tibble: 80 x 2
#>    product        value
#>    <chr>          <dbl>
#>  1 Singapore     0.177 
#>  2 South Africa  0.0653
#>  3 Lebanon      -0.0497
#>  4 Seychelles   -0.0971
#>  5 Philippines  -0.163 
#>  6 Morocco      -0.170 
#>  7 Tunisia      -0.170 
#>  8 Angola       -0.173 
#>  9 Aruba        -0.173 
#> 10 Bahamas      -0.173 
#> # … with 70 more rows
#> 
#> $complexity_outlook_gain
#> # A tibble: 880 x 3
#>    country    product             value
#>    <chr>      <chr>               <dbl>
#>  1 Albania    1 Transportation -0.0213 
#>  2 Algeria    1 Transportation -0.00188
#>  3 Angola     1 Transportation -0.00691
#>  4 Argentina  1 Transportation -0.0127 
#>  5 Armenia    1 Transportation -0.0405 
#>  6 Aruba      1 Transportation -0.00691
#>  7 Australia  1 Transportation -0.0213 
#>  8 Azerbaijan 1 Transportation  0.0207 
#>  9 Bahamas    1 Transportation -0.00691
#> 10 Bangladesh 1 Transportation -0.0221 
#> # … with 870 more rows
#>