Countries position

Countries position

ec_countries_position(rca = NULL, c1 = "country", p1 = "product",
  v1 = "value", pp = NULL, p21 = "from", p22 = "to",
  v2 = "value", pci = NULL, p3 = "product", v3 = "value",
  tbl = FALSE)

Arguments

rca
matrix or tibble/data.frame (e.g. the output of revealed_comparative_advantage()). If the input is a matrix it must be a zero/one matrix with countries in rows and products in columns. If the input is a tibble/data.frame it must contain at least three columns with countries, products and values.

c1
string to indicate the column that contains exporting countries in rca (set to "country" by default)

p1
string to indicate the column that contains exported products in rca (set to "product" by default)

v1
string to indicate the column that contains traded values in rca (set to "value" by default)

pp
matrix or tibble/data.frame (e.g. the output of proximity_matrices()). If the input is a matrix it must be a numeric matrix with products in both rows and columns. If the input is a tibble/data.frame it must contain at least three columns columns with products (twice) and values.

p21
string to indicate the first column that contains exported products in pp (set to "from" by default)

p22
string to indicate the second column that contains exported products in pp (set to "to" by default)

v2
string to indicate the column that contains proximity values in pp (set to "value" by default)

pci
numeric vector or tibble/data.frame with the product complexity index (e.g. the product index from ec_complexity_measures()). If the input is a vector it must be numeric with optional names. If the input is a tibble/data.frame it must contain at least two columns with products and values.

p3
string to indicate the column that contains exported products in pp (e.g. "product")

v3
string to indicate the column that contains proximity values in pp (e.g. "value") from ec_complexity_measures())

tbl
logical value to use tibble output instead of a matrix output (set to FALSE by default)

References

For more information on proximity distance, complexity outlook, complexity outlook gain and its applications see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

Examples

ec_countries_position(
  rca = ec_output_demo$rca_tbl,
  c1 = "country",
  p1 = "product",
  v1 = "value",
  pp = ec_output_demo$proximity_tbl$proximity_p,
  p21 = "from",
  p22 = "to",
  v2 = "value",
  pci = ec_output_demo$complexity_measures_tbl$complexity_index_p,
  p3 = "product",
  v3 = "value",
  tbl = TRUE
)

#> Joining, by = c("from", "to")

#> $proximity_distance
#> # A tibble: 156,578 x 3
#>    country product value
#>    <chr>   <chr>   <dbl>
#>  1 afg     0101    0.960
#>  2 ago     0101    0.948
#>  3 alb     0101    0.955
#>  4 and     0101    0.974
#>  5 ant     0101    0.980
#>  6 are     0101    1    
#>  7 arg     0101    0.861
#>  8 aus     0101    0.896
#>  9 aut     0101    0.609
#> 10 bdi     0101    0.989
#> # ... with 156,568 more rows
#> 
#> $complexity_outlook
#> # A tibble: 158 x 2
#>    product value
#>    <chr>   <dbl>
#>  1 nld      268.
#>  2 bel      265.
#>  3 ita      259.
#>  4 fra      254.
#>  5 gbr      244.
#>  6 che      242.
#>  7 jpn      239.
#>  8 aut      236.
#>  9 swe      217.
#> 10 csk      214.
#> # ... with 148 more rows
#> 
#> $complexity_outlook_gain
#> # A tibble: 156,578 x 3
#>    country product       value
#>    <chr>   <chr>         <dbl>
#>  1 afg     0101    -0.0000323 
#>  2 ago     0101    -0.0000131 
#>  3 alb     0101    -0.0000418 
#>  4 and     0101    -0.0000139 
#>  5 ant     0101    -0.00000536
#>  6 are     0101     0.00000165
#>  7 arg     0101    -0.000154  
#>  8 aus     0101    -0.0000690 
#>  9 aut     0101    -0.000299  
#> 10 bdi     0101     0.00000546
#> # ... with 156,568 more rows
#>