Networks

Networks

ec_networks(pc, pp, cutoff_c = 0.2, cutoff_p = 0.4, tbl = FALSE,
  compute = "both")

Arguments

pc
matrix or tibble/data.frame, if d is a tibble/ data.frame it must contain the columns from (character/factor), to (character/factor) and value (numeric), if it is a matrix it must be a numeric matrix with countries in row names and column names

pp
matrix or tibble/data.frame, if d is a tibble/data.frame it must contain the columns from (character/factor), to (character/factor) and value (numeric), if it is a matrix it must be a numeric matrix with products in row names and column names

cutoff_c
all the values below the specified cutoff_c will be converted to 0 and excluded from the countries network (default set to 0.2)

cutoff_p
all the values below the specified cutoff_p will be converted to 0 and excluded from the products network (default set to 0.4)

tbl
when set to TRUE the output will be a tibble instead of a graph (default set to FALSE)

compute
by default set to "both", it can also be "country" or "product"

References

For more information on networks such as the product space and its applications see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

Examples

ec_networks(
  pc = ec_output_demo$proximity_tbl$proximity_c,
  pp = ec_output_demo$proximity_tbl$proximity_p,
  tbl = TRUE
)

#> $network_c
#> # A tibble: 497 x 3
#>    from  to    value
#>    <chr> <chr> <dbl>
#>  1 khm   cmr   0.2  
#>  2 khm   ecu   0.2  
#>  3 khm   gtm   0.225
#>  4 khm   idn   0.214
#>  5 khm   mdg   0.204
#>  6 khm   mli   0.325
#>  7 khm   nic   0.2  
#>  8 khm   pak   0.241
#>  9 khm   sdn   0.2  
#> 10 khm   nga   0.267
#> # ... with 487 more rows
#> 
#> $network_p
#> # A tibble: 56,160 x 3
#>    from  to    value
#>    <chr> <chr> <dbl>
#>  1 0711  9614  0.467
#>  2 0711  0103  0.4  
#>  3 0711  0203  0.467
#>  4 0711  0207  0.4  
#>  5 0711  0210  0.4  
#>  6 0711  0405  0.4  
#>  7 0711  0407  0.5  
#>  8 0711  0505  0.467
#>  9 0711  0701  0.529
#> 10 0711  0702  0.467
#> # ... with 56,150 more rows
#>