Example outputs of the functions within the package

List of different economic complexity computations obtained by using the functions within the package. The output is provided as a list of tibbles.

ec_output_demo

Format

The format is: List of 3 $ rca_tbl :Classes ‘tbl_df’, ‘tbl’ and 'data.frame': 44435 obs. of 3 variables: ..$ country: chr [1:44435] "afg" "afg" "afg" "afg" ... ..$ product: chr [1:44435] "0106" "0405" "0407" "0501" ... ..$ value : num [1:44435] 0 0 0 1 1 0 0 1 1 1 ... $ complexity_measures_tbl:List of 4 ..$ complexity_index_c:Classes ‘tbl_df’, ‘tbl’ and 'data.frame': 99 obs. of 2 variables: .. ..$ country: chr [1:99] "deu" "fra" "usa" "gbr" ... .. ..$ value : num [1:99] 19.47 11.99 11.55 8.43 6.92 ... ..$ complexity_index_p:Classes ‘tbl_df’, ‘tbl’ and 'data.frame': 991 obs. of 2 variables: .. ..$ product: chr [1:991] "2702" "7111" "2850" "8804" ... .. ..$ value : num [1:991] 48.1 48.1 29.7 18.4 17.9 ... ..$ diversity :Classes ‘tbl_df’, ‘tbl’ and 'data.frame': 99 obs. of 2 variables: .. ..$ country: chr [1:99] "ago" "alb" "are" "arg" ... .. ..$ value : num [1:99] 52 32 1 83 81 290 320 127 27 70 ... ..$ ubiquity :Classes ‘tbl_df’, ‘tbl’ and 'data.frame': 991 obs. of 2 variables: .. ..$ product: chr [1:991] "0101" "0102" "0103" "0104" ... .. ..$ value : num [1:991] 13 19 13 12 9 27 17 12 4 10 ... $ proximity_tbl :List of 2 ..$ proximity_c:Classes ‘tbl_df’, ‘tbl’ and 'data.frame': 4484 obs. of 3 variables: .. ..$ from : chr [1:4484] "alb" "are" "arg" "aus" ... .. ..$ to : chr [1:4484] "ago" "ago" "ago" "ago" ... .. ..$ value: num [1:4484] 0.0769 0.0192 0.1325 0.0741 0.0241 ... ..$ proximity_p:Classes ‘tbl_df’, ‘tbl’ and 'data.frame': 408869 obs. of 3 variables: .. ..$ from : chr [1:408869] "0102" "0103" "0104" "0106" ... .. ..$ to : chr [1:408869] "0101" "0101" "0101" "0101" ... .. ..$ value: num [1:408869] 0.316 0.308 0.385 0.148 0.353 ...

Examples

ec_output_demo$rca_tbl

#> # A tibble: 44,435 x 3
#>    country product value
#>    <chr>   <chr>   <dbl>
#>  1 afg     0106        0
#>  2 afg     0405        0
#>  3 afg     0407        0
#>  4 afg     0501        1
#>  5 afg     0504        1
#>  6 afg     0703        0
#>  7 afg     0712        0
#>  8 afg     0714        1
#>  9 afg     0801        1
#> 10 afg     0802        1
#> # … with 44,425 more rows

ec_output_demo$complexity_measures_tbl$complexity_index_c

#> # A tibble: 99 x 2
#>    country value
#>    <chr>   <dbl>
#>  1 deu     19.5 
#>  2 fra     12.0 
#>  3 usa     11.6 
#>  4 gbr      8.43
#>  5 che      6.92
#>  6 nld      6.63
#>  7 ita      6.55
#>  8 swe      4.85
#>  9 bel      4.32
#> 10 aut      4.29
#> # … with 89 more rows