Productivity Levels

Productivity Levels

ec_productivity_levels(data = NULL, c1 = "country", p1 = "product",
  v1 = "value", gdp = NULL, c2 = "country", v2 = "value",
  tbl = FALSE)

Arguments

data
matrix or tibble/data.frame (e.g. world_trade_2017). If the input is a matrix it must be a zero/one matrix with countries in rows and products in columns. If the input is a tibble/data.frame it must contain at least three columns with countries, products and values.

c1
string to indicate the column that contains exporting countries in revealed_comparative_advantage (set to "country" by default)

p1
string to indicate the column that contains exported products in revealed_comparative_advantage (set to "product" by default)

v1
string to indicate the column that contains traded values in revealed_comparative_advantage (set to "value" by default)

gdp
vector or tibble/data.frame (e.g. world_gdp_and_population_2017). If the input is a vector it must be a numeric vector with optional names. If the input is a tibble/data.frame it must contain at least two columns with countries and values.

c2
string to indicate the column that contains exporting countries in revealed_comparative_advantage (set to "country" by default)

v2
string to indicate the column that contains traded values in revealed_comparative_advantage (set to "value" by default)

tbl
when set to TRUE the output will be a tibble instead of a matrix (default set to FALSE)

References

For more information on prody and its applications see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

Hausmann R, Hwang J, Rodrik D (2005). “What You Export Matters.” Working Paper 11905, National Bureau of Economic Research. doi: 10.3386/w11905 .

Examples

ec_productivity_levels(
  data = ec_trade_1962,
  c1 = "country",
  p1 = "product",
  v1 = "value",
  gdp = ec_gdp_pc_1962,
  c2 = "country",
  v2 = "value",
  tbl = TRUE
)

#> Warning: Joining data and gdp resulted in a table with less
#>     reporting countries than those in data.

#> Warning: Joining data and gdp resulted in a table with less
#>     reporting countries than those in gdp.

#> $productivity_level_country
#> # A tibble: 89 x 2
#>    name  value
#>    <chr> <dbl>
#>  1 afg   1044.
#>  2 arg   1097.
#>  3 aus   1556.
#>  4 aut   1035.
#>  5 bdi   1010.
#>  6 bel    981.
#>  7 ben   1036.
#>  8 bfa    939.
#>  9 bhs   1037.
#> 10 blz   1747.
#> # … with 79 more rows
#> 
#> $productivity_level_product
#> # A tibble: 991 x 2
#>    name  value
#>    <chr> <dbl>
#>  1 0101   816.
#>  2 0102   359.
#>  3 0103   217.
#>  4 0104   184.
#>  5 0105   223.
#>  6 0106   441.
#>  7 0201   909.
#>  8 0203  1313.
#>  9 0204  2141.
#> 10 0205   993.
#> # … with 981 more rows
#>