Proximity

Proximity

ec_proximity(rca = NULL, c = "country", p = "product", v = "value",
  d = NULL, d_c = "country", d_v = "value", u = NULL,
  u_p = "product", u_v = "value", tbl = FALSE, compute = "both")

Arguments

rca
matrix or tibble/data.frame in long format (e.g. the output of revealed_comparative_advantage()). If it is a matrix it must be a zero/one matrix with countries in the row names and ps in the column names. If rca is a tibble/data.frame it must contain the columns c (character/ factor), p (character/factor) and discrete RCA (integer)

c
string to indicate the column that contains exporting countries (default set to "country" that is the output of revealed_comparative_advantage(), applies only if rca is a data.frame)

p
string to indicate the column that contains exported ps (default set to "product" that is the output of revealed_comparative_advantage(), applies only if rca is a data.frame)

v
string to indicate the column that contains RCA vs (default set to "value" that is the output of revealed_comparative_advantage(), applies only if rca is a data.frame)

d
numeric vector or tibble/data.frame containing diversity measures (e.g. d from complexity_measures())

d_c
string to indicate the column that contains countries in "diversity" (default set to "country" that is the output of economic_complexity_measures())

d_v
string to indicate the column that contains values in "diversity" (default set to "value" that is the output of economic_complexity_measures())

u
numeric vector or tibble/data.frame containing ubiquity measures (e.g. u from complexity_measures())

u_p
string to indicate the column that contains products in "ubiquity" (default set to "product" that is the output of economic_complexity_measures())

u_v
string to indicate the column that contains values in "ubiquity" (default set to "value" that is the output of economic_complexity_measures())

tbl
when set to TRUE the output will be a tibble instead of a matrix (default set to FALSE)

compute
by default set to "both", it can also be "country" or "product"

References

For more information on proximity and its applications see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

Examples

ec_proximity(
  rca = ec_output_demo$rca_tbl,
  d = ec_output_demo$complexity_measures_tbl$diversity,
  u = ec_output_demo$complexity_measures_tbl$ubiquity,
  tbl = TRUE
)

#> $proximity_c
#> # A tibble: 4,484 x 3
#>    from  to     value
#>    <chr> <chr>  <dbl>
#>  1 alb   ago   0.0769
#>  2 are   ago   0.0192
#>  3 arg   ago   0.133 
#>  4 aus   ago   0.0741
#>  5 aut   ago   0.0241
#>  6 bel   ago   0.0281
#>  7 bgr   ago   0.110 
#>  8 bol   ago   0.0385
#>  9 bra   ago   0.2   
#> 10 can   ago   0.0845
#> # … with 4,474 more rows
#> 
#> $proximity_p
#> # A tibble: 408,869 x 3
#>    from  to    value
#>    <chr> <chr> <dbl>
#>  1 0102  0101  0.316
#>  2 0103  0101  0.308
#>  3 0104  0101  0.385
#>  4 0106  0101  0.148
#>  5 0201  0101  0.353
#>  6 0203  0101  0.462
#>  7 0204  0101  0.154
#>  8 0205  0101  0.385
#>  9 0206  0101  0.385
#> 10 0207  0101  0.429
#> # … with 408,859 more rows
#>