economiccomplexity: Computational Methods for Economic Complexity

A wrapper of different methods from Linear Algebra for the equations introduced in The Atlas of Economic Complexity and related literature. This package provides standard matrix and graph output that can be used seamlessly with other packages. See doi:10.21105/joss.01866 for a summary of these methods and its evolution in literature.

See also

Author

Maintainer: Mauricio Vargas Sepulveda m.sepulveda@mail.utoronto.ca (ORCID) [copyright holder]

Other contributors:

Carlo Bottai (improved the eigenvalues calculation) [contributor]

Diego Kozlowski (provided initial RCA function used up to v0.2.4) [contributor]

Nico Pintar (suggested improvements to special cases in eigenvalues calculation) [reviewer]

The World Bank (World GDP per capita dataset) [data contributor]

Open Trade Statistics (World Trade dataset) [data contributor]