Example outputs of the functions within the package

List of different economic complexity computations obtained by using the functions within the package. The output is provided as a list of tibbles.

economiccomplexity_output

Format

The format is: List of 3 $ rca_tbl :Classes ‘tbl_df’, ‘tbl’ and 'data.frame': 44435 obs. of 3 variables: ..$ country: chr [1:44435] "afg" "afg" "afg" "afg" ... ..$ product: chr [1:44435] "0106" "0405" "0407" "0501" ... ..$ value : num [1:44435] 0 0 0 1 1 0 0 1 1 1 ... $ complexity_measures_tbl:List of 4 ..$ complexity_index_c:Classes ‘tbl_df’, ‘tbl’ and 'data.frame': 99 obs. of 2 variables: .. ..$ country: chr [1:99] "deu" "fra" "usa" "gbr" ... .. ..$ value : num [1:99] 19.47 11.99 11.55 8.43 6.92 ... ..$ complexity_index_p:Classes ‘tbl_df’, ‘tbl’ and 'data.frame': 991 obs. of 2 variables: .. ..$ product: chr [1:991] "2702" "7111" "2850" "8804" ... .. ..$ value : num [1:991] 48.1 48.1 29.7 18.4 17.9 ... ..$ diversity :Classes ‘tbl_df’, ‘tbl’ and 'data.frame': 99 obs. of 2 variables: .. ..$ country: chr [1:99] "ago" "alb" "are" "arg" ... .. ..$ value : num [1:99] 52 32 1 83 81 290 320 127 27 70 ... ..$ ubiquity :Classes ‘tbl_df’, ‘tbl’ and 'data.frame': 991 obs. of 2 variables: .. ..$ product: chr [1:991] "0101" "0102" "0103" "0104" ... .. ..$ value : num [1:991] 13 19 13 12 9 27 17 12 4 10 ... $ proximity_tbl :List of 2 ..$ proximity_c:Classes ‘tbl_df’, ‘tbl’ and 'data.frame': 4484 obs. of 3 variables: .. ..$ from : chr [1:4484] "alb" "are" "arg" "aus" ... .. ..$ to : chr [1:4484] "ago" "ago" "ago" "ago" ... .. ..$ value: num [1:4484] 0.0769 0.0192 0.1325 0.0741 0.0241 ... ..$ proximity_p:Classes ‘tbl_df’, ‘tbl’ and 'data.frame': 408869 obs. of 3 variables: .. ..$ from : chr [1:408869] "0102" "0103" "0104" "0106" ... .. ..$ to : chr [1:408869] "0101" "0101" "0101" "0101" ... .. ..$ value: num [1:408869] 0.316 0.308 0.385 0.148 0.353 ...

Examples

economiccomplexity_output$balassa_index

#> 9 x 12 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 12 column names ‘alpha’, ‘beta’, ‘delta’ ... ]]

#>                                                    
#> Bepis 9                     . . . 1 . . . . . . 1 .
#> Bird World                  . . . . . 1 . . . . . .
#> Crepktct 4                  . . . . . 1 . 1 1 . . .
#> Earth C-137                 1 . 1 1 . . . . 1 1 1 1
#> Galactic Federation Outpost . . . 1 . 1 . 1 . . . .
#> Galactic Federation Prison  . 1 . . . 1 . . . . . 1
#> Gear World                  . . . 1 1 . . . . 1 1 .
#> Glaagablaaga                . . . . 1 1 1 1 . 1 . .
#> Gromflom Prime              . 1 . . . . . . . . . .

economiccomplexity_output$complexity_measures$complexity_index_x

#> NULL