Networks

Networks

networks(proximity_countries, proximity_products, countries_cutoff = 0.2,
  products_cutoff = 0.4, tbl_output = FALSE, compute = "both")

Arguments

proximity_countries
matrix or tibble/data.frame, if d is a tibble/ data.frame it must contain the columns from (character/factor), to (character/factor) and value (numeric), if it is a matrix it must be a numeric matrix with countries in row names and column names

proximity_products
matrix or tibble/data.frame, if d is a tibble/data.frame it must contain the columns from (character/factor), to (character/factor) and value (numeric), if it is a matrix it must be a numeric matrix with products in row names and column names

countries_cutoff
all the values lower than the specified countries_cutoff will be converted to 0 and excluded from the countries network (default set to 0.2)

products_cutoff
all the values lower than the specified products_cutoff will be converted to 0 and excluded from the products network (default set to 0.4)

tbl_output
when set to TRUE the output will be a tibble instead of a graph (default set to FALSE)

compute
by default set to "both", it can also be "country" or "product"

References

For more information on networks such as the product space and its applications see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

Examples

networks(
 proximity_countries =
  package_output_demo$proximity_matrix$proximity_countries,
 proximity_products =
  package_output_demo$proximity_matrix$proximity_products,
 tbl_output = TRUE
)

#> $network_countries
#> # A tibble: 3,063 x 3
#>    from       to         value
#>    <chr>      <chr>      <dbl>
#>  1 Cabo Verde Namibia    1    
#>  2 Cabo Verde Algeria    0.2  
#>  3 Cabo Verde Angola     0.333
#>  4 Cabo Verde Anguilla   0.333
#>  5 Cabo Verde Argentina  0.25 
#>  6 Cabo Verde Armenia    0.5  
#>  7 Cabo Verde Aruba      0.333
#>  8 Cabo Verde Australia  0.667
#>  9 Cabo Verde Azerbaijan 0.333
#> 10 Cabo Verde Bahamas    0.333
#> # … with 3,053 more rows
#> 
#> $network_products
#> NULL
#>