Productivity Levels

productivity_levels() computes EXPY and PRODY measures.

productivity_levels(data_exp, data_gdp, country = "country",
  product = "product", value = "value")

Arguments

data_exp
(Type: data.frame, matrix or dgCMatrix) a dataset containing countries, products and exported values.

data_gdp
(Type: data.frame, matrix or dgCMatrix) a dataset containing countries and per-capita GDP values.

country
(Type: character) the column with the countries. By default this is set to "country". Used only if the input is a data.frame.

product
(Type: character) the column with the products. By default this is set to "product". Used only if the input is a data.frame.

value
(Type: character) the column with the metric for country-product pairs. By default this is set to "value". Used only if the input is a data.frame.

Value

A list of two named numeric vectors.

Details

The current implementation follows (Hausmann et al. 2005) to obtain the productivity and income levels associated to exports.

References

For more information on prody and its applications see:

Hausmann R, Hwang J, Rodrik D (2005). “What You Export Matters.” Working Paper 11905, National Bureau of Economic Research. doi: 10.3386/w11905 .

Examples

data_exp <- galactic_federation

# Since I don't have a per-capita GDP dataset for the Galactic Federation,
# I'll create simulated data
set.seed(1810)
data_gdp <- setNames(
 rnorm(1:nrow(galactic_federation), 1000, 200),
 rownames(galactic_federation)
)

productivity_levels(data_exp, data_gdp)

#> Warning: 'data_exp' and 'data_gdp' don't have the same countries, some elements will be dropped

#> $productivity_level_country
#>                     Bepis 9                  Bird World 
#>                   1074.1218                   1081.7712 
#>                  Crepktct 4                 Earth C-137 
#>                   1090.0985                   1147.3059 
#> Galactic Federation Outpost  Galactic Federation Prison 
#>                   1072.8671                   1062.9803 
#>                  Gear World                Glaagablaaga 
#>                    974.1316                   1144.2559 
#>              Gromflom Prime 
#>                    990.1308 
#> 
#> $productivity_level_product
#>     alpha      beta     delta   epsilon       eta     gamma      iota     kappa 
#> 1305.3414  979.0985 1290.8110 1066.3034  879.5715 1081.7712 1274.9481 1133.4150 
#>    lambda        mu     theta      zeta 
#> 1316.7840  909.5411 1032.4238 1167.3903 
#>