Revealed Comparative Advantage (RCA)

Revealed Comparative Advantage (RCA)

revealed_comparative_advantage(trade_data = NULL, country = "country",
  product = "product", value = "value", cutoff = 1,
  discrete = TRUE, tbl_output = FALSE)

Arguments

trade_data
tibble/data.frame in long format, it must contain the columns country (character/factor), product (character/factor) and export value (numeric)

country
string to indicate the column that contains exporting countries (e.g. "reporter_iso")

product
string to indicate the column that contains exported products (e.g. "product_code")

value
string to indicate the column that contains traded values (e.g. "trade_value_usd")

cutoff
when set to TRUE all the values lower than the specified cutoff will be converted to zero and to one in other case, numeric (default set to 1)

discrete
when set to TRUE it will convert all the Revealed Comparative Advantage values to zero or one based on the cutoff value (default set to TRUE)

tbl_output
when set to TRUE the output will be a tibble instead of a matrix (default set to FALSE)

References

For more information on revealed comparative advantage and its uses see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

Examples

revealed_comparative_advantage(
 trade_data = services_trade_2016$services_trade_2016_matrix,
 tbl_output = TRUE
)

#> # A tibble: 735 x 3
#>    country product                                          value
#>    <chr>   <chr>                                            <dbl>
#>  1 Albania 1 Transportation                                     0
#>  2 Albania 10 Personal, cultural, and recreational services     1
#>  3 Albania 11 Government services, n.i.e.                       0
#>  4 Albania 2 Travel                                             1
#>  5 Albania 3 Communications services                            1
#>  6 Albania 4 Construction services                              0
#>  7 Albania 5 Insurance services                                 0
#>  8 Albania 6 Financial services                                 0
#>  9 Albania 7 Computer and information services                  0
#> 10 Albania 8 Royalties and license fees                         0
#> # … with 725 more rows