Exported services in 2016 (matrix and tibble version)

Example dataset with countries and the services they export. The exports are grouped to the most aggregated categories of EBOPS2002 service codes and product data is not included to keep the example lightweight.

services_trade_2016

Format

The format is: List of 2 $ services_trade_2016_matrix:Formal class 'dgCMatrix' [package "Matrix"] with 6 slots .. ..@ i : int [1:736] 0 1 2 3 4 5 6 7 8 9 ... .. ..@ p : int [1:12] 0 80 145 218 298 355 409 475 540 601 ... .. ..@ Dim : int [1:2] 80 11 .. ..@ Dimnames:List of 2 .. .. ..$ : chr [1:80] "Albania" "Algeria" "Angola" "Argentina" ... .. .. ..$ : chr [1:11] "1 Transportation" "10 Personal, cultural, and recreational services" "11 Government services, n.i.e." "2 Travel" ... .. ..@ x : num [1:736] 2.00e+08 6.72e+08 3.26e+07 2.18e+09 1.88e+08 ... .. ..@ factors : list() $ services_trade_2016_tibble:Classes ‘spec_tbl_df’, ‘tbl_df’, ‘tbl’ and 'data.frame': 736 obs. of 3 variables: ..$ country: chr [1:736] "Albania" "Algeria" "Angola" "Azerbaijan" ... ..$ product: chr [1:736] "1 Transportation" "1 Transportation" "1 Transportation" "1 Transportation" ... ..$ value : num [1:736] 2.00e+08 6.72e+08 3.26e+07 1.10e+09 2.18e+09 ... ..- attr(*, "spec")=List of 3 .. ..$ cols :List of 24 .. .. ..$ Classification : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_character" "collector" .. .. ..$ Year : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector" .. .. ..$ Period : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector" .. .. ..$ Period Desc. : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector" .. .. ..$ Aggregate Level : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector" .. .. ..$ Is Leaf Code : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector" .. .. ..$ Trade Flow Code : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector" .. .. ..$ Trade Flow : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_character" "collector" .. .. ..$ Reporter Code : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector" .. .. ..$ Reporter : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_character" "collector" .. .. ..$ Reporter ISO : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_logical" "collector" .. .. ..$ Partner Code : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector" .. .. ..$ Partner : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_character" "collector" .. .. ..$ Partner ISO : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_logical" "collector" .. .. ..$ Mode of Transport Code: list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_logical" "collector" .. .. ..$ Mode of Transport : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_logical" "collector" .. .. ..$ Commodity Code : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector" .. .. ..$ Commodity : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_character" "collector" .. .. ..$ Netweight (kg) : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_logical" "collector" .. .. ..$ Gross weight (kg) : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_logical" "collector" .. .. ..$ Trade Value (US$) : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector" .. .. ..$ CIF Trade Value (US$) : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_logical" "collector" .. .. ..$ FOB Trade Value (US$) : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_logical" "collector" .. .. ..$ Flag : list() .. .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector" .. ..$ default: list() .. .. ..- attr(*, "class")= chr [1:2] "collector_guess" "collector" .. ..$ skip : num 1 .. ..- attr(*, "class")= chr "col_spec"

Source

United Nations COMTRADE

Examples

services_trade_2016$services_trade_2016_matrix

#> 80 x 11 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 11 column names ‘1 Transportation’, ‘10 Personal, cultural, and recreational services’, ‘11 Government services, n.i.e.’ ... ]]

#>                                                                        
#> Albania                   200490000   18140000    40270000   1692860000
#> Algeria                   672372922    5147535   289928047    209347434
#> Angola                     32634214          .           .    622816341
#> Argentina                2175160000  232460000   194830000   4686630000
#> Armenia                   188458900   18740000    32076919    967701004
#> Aruba                     107912730    3607604    29210118   1630842603
#> Australia                4340600000  620000000   763400000  32423100000
#> Azerbaijan               1095425000   15199000    27336000   2713789000
#> Bahamas                    95553064          .    36413144   2590458326
#> Bangladesh                412300000     500000  1232600000    213100000
#> Belarus                  2902900000   13400000    26100000    709600000
#> Belize                     27003777          .     5142809    390873899
#> Bhutan                   2962863685          .   113866542   6485259329
#> Bolivia                   291998880    1693776    24654161    713059512
#> Bosnia Herzegovina        350644591    1030498     1881965    724075385
#> Botswana                   33181970     -91815    76218930    578763253
#> Brazil                   5057814902  567561443   732588042   6023798709
#> Brunei Darussalam         345716252          .    21725480    144152245
#> Burundi                     2198970      32434    55273350      1886471
#> Cambodia                  456077728    1188672   185471962   3212304014
#> Cameroon                  484700000   87700000    83200000    505300000
#> Chile                    2952887944   25214389   356163635   2737106041
#> China                   33827133238  742330878  1211714453  44431583511
#> China, Macao SAR          588740563    4213112           .  30507427419
#> Colombia                 1630935933  129632064   135926840   4522459950
#> Congo                      63938026     104093    43143989      4256867
#> Costa Rica                418902445     150000    29024758   3878813802
#> Dem. Rep. of the Congo     63938026     104093    43143989      4256867
#> Djibouti                  121370695          .   221837722     66959262
#> Dominica                     688697          .      951656    235166688
#> Dominican Rep.            581000000   43200000   320700000   6719600000
#> Ecuador                   409805398   42848233   136478895   1443595948
#> El Salvador               445638141     358500    59700000    829299820
#> Georgia                   915454022   15347108    94559221   2110709408
#> Greece                   8618389341  235597269   170860584  14614776948
#> Grenada                    14143696          .     2554543    437312333
#> Indonesia                3566993504   81115124   678202656  12203043823
#> Israel                   3580000000  392680000    15000000   5883000000
#> Japan                   31651150981  811961825  5087724629  30751055596
#> Jordan                   1249224260   59379408   191396333   4049224260
#> Kuwait                   1231120479          .   561821377    600443840
#> Kyrgyzstan                159373044     503418     6581576    432364315
#> Lebanon                   944000000  434000000   293000000   7044000000
#> Lesotho                      500000          .     2470000     46710000
#> Madagascar                242511794    1838017    24037608    750134183
#> Mauritius                 356831785   24450940    29768808   1571922516
#> Mexico                   1598427447   13445000   146870932  19649652001
#> Mongolia                  255424979    1594956     2978661    316285942
#> Morocco                  2573331390   69353149   698898525   6553336870
#> Mozambique                264006507          .           .    107851512
#> Myanmar                   287888020   71445924    87588291   2197150000
#> Namibia                    78100130    2463417    48652687    309710863
#> New Zealand              1849901697  287125981   169062013   9672163519
#> Philippines              1896758094  169065172    17442971   5142672850
#> Poland                  13400310440  700326330           .  11005556610
#> Qatar                    7607142857  378846154   625824176   5410989011
#> Rep. of Korea           26437000000 1130800000  1027600000  17210300000
#> Rep. of Moldova           338700000    2090000    36770000    254800000
#> Russian Federation      18694087468  468243295   916373281   8653351388
#> Samoa                       6375955    2484088           .    151455196
#> Sao Tome and Principe        285259     414409     4862709     68859083
#> Saudi Arabia             3084266667          .   690144000  11095566274
#> Serbia                   1030000000  158000000     9000000   1040000000
#> Seychelles                203300528          .    15037437    413669222
#> Singapore               42861981634  831129584   273303606  18944297134
#> Slovenia                 2031535240   99221230    16110950   2433753440
#> Solomon Isds               33442167    1291392     7549818     56235441
#> South Africa             2191761880  197774503   387769056   7940803246
#> Sri Lanka                2250146027          .    33218215   3518489413
#> Suriname                   32854556     331195    10861617     63529153
#> Tajikistan                161103590          .           .      3604600
#> TFYR of Macedonia         356759058   31056364     3624126    279831490
#> Thailand                 5724070000   60060000   290890000  48792367579
#> Trinidad and Tobago       337404354       2618    13249946    463502706
#> Tunisia                   948628382   10013507   180708863   1236644777
#> Uganda                    128867512   15384864   243614481   1060348380
#> United Rep. of Tanzania   128867512   15384864   243614481   1060348380
#> Uruguay                   441861543   58743136    63478256   2071155162
#> USA                     84318000000 3275000000 18814000000 205940000000
#> Zambia                     38341170   35500000           .    682591895
#>                                                                        
#> Albania                   120550000     8310000     7520000     2130000
#> Algeria                           .   202285882    88254508           .
#> Angola                            .           .           .           .
#> Argentina                 172020000     7280000    31930000    53530000
#> Armenia                    31079700   149628765    20597007     6141616
#> Aruba                             .     4647725           .    16853941
#> Australia                1358900000   130100000   392500000  2627100000
#> Azerbaijan                        .    31058000    12541000           .
#> Bahamas                           .           .           .           .
#> Bangladesh                447800000   103000000      800000    89200000
#> Belarus                   210000000  1031200000    17800000    29400000
#> Belize                     11846032           .      320102     3166574
#> Bhutan                     45351351           .   240688329   174203238
#> Bolivia                    42867749      724029     4091994     3642940
#> Bosnia Herzegovina         90122665    66531389     1755379     1702288
#> Botswana                          .    63624511     2125977     1791751
#> Brazil                            .   147668765   783649700           .
#> Brunei Darussalam                 .           .     1404265           .
#> Burundi                     3101860           .     1254889     1182110
#> Cambodia                          .    13076239      404812           .
#> Cameroon                          .     1200000    57900000    38700000
#> Chile                             .           .   307109717   174317123
#> China                             . 12660612351  4063790929  3174183105
#> China, Macao SAR           37231601           .   434732758   276614098
#> Colombia                  203679319           .    22811110    31014463
#> Congo                       5795559           .           .     3881679
#> Costa Rica                 26995406           .     1668456    60809600
#> Dem. Rep. of the Congo            .           .           .           .
#> Djibouti                          .           .           .           .
#> Dominica                          .           .     2660099           .
#> Dominican Rep.             99900000           .    47700000    31900000
#> Ecuador                    66088747           .           .    41032425
#> El Salvador               123083322    11620000    47104478    45751327
#> Georgia                    34008373     9813789    13502914    13130097
#> Greece                    457251252   552419712   334970847   112431576
#> Grenada                      587599           .     6456699      444936
#> Indonesia                 759700664   237848691    52580570   262038188
#> Israel                    462680000   770300000    36000000   655050000
#> Japan                             .  9379331534  1702014795           .
#> Jordan                            .    64174894           .    96755994
#> Kuwait                   2567417510           .   446858688   212587273
#> Kyrgyzstan                 62547959    53875636      156219     4592743
#> Lebanon                   702000000   291000000   220000000  2401000000
#> Lesotho                     3320000           .       10000       10000
#> Madagascar                 20044553    26839848     3260440      386484
#> Mauritius                  82722398    18232692    24422803    53066137
#> Mexico                     99909000           .  2502390351   146194934
#> Mongolia                   14245012    27054269           .    15815183
#> Morocco                   853685529   534652882    99351421    63669803
#> Mozambique                 24242533           .    20650961      525012
#> Myanmar                           .    63670256           .    26804059
#> Namibia                    17663429     6116960      418663     8756451
#> New Zealand                45409218    23053911   108283521   511377660
#> Philippines               310011398    71285135    76659881   142233718
#> Poland                    641771360  1745649210   463217590   563772140
#> Qatar                     634340659           .   518406593           .
#> Rep. of Korea                     . 10952800000   620200000  1768200000
#> Rep. of Moldova            84620000     3280000      150000     5070000
#> Russian Federation       1639082220  3952569463   426529790  1300377118
#> Samoa                      17579697    14007603      483559    12775308
#> Sao Tome and Principe       1369903     1639241           .       17085
#> Saudi Arabia                      .           .   221066667   573569732
#> Serbia                    151000000   180000000    22000000    24000000
#> Seychelles                 12086888           .           .           .
#> Singapore                1881278425  1033791157  5879646490 16446565633
#> Slovenia                  400440440   428217940   103443410    28555270
#> Solomon Isds                4338537     2036395       44492     1210212
#> South Africa                      .    16993047           .   998153856
#> Sri Lanka                         .    63166072   121001430   251780146
#> Suriname                    6493929     1279695    16054744     1373023
#> Tajikistan                 10178610       44760           .           .
#> TFYR of Macedonia          61938721   160591724       50594     2839361
#> Thailand                          .   462700000    68210000   724000000
#> Trinidad and Tobago        11661925     1791596   125647019      576202
#> Tunisia                   273997485   302640771    60686507    62828932
#> Uganda                     17964762   103884705    38779487    26655026
#> United Rep. of Tanzania    17964761   103884705    38779487    26655026
#> Uruguay                   190830169           .     3086919   132031404
#> USA                     12225000000  1404000000 16348000000 98180000000
#> Zambia                     34924141           .    88912612           .
#>                                                              
#> Albania                     6810000      2700000    202870000
#> Algeria                           .       475375   1486866665
#> Angola                     21658540     10997630     22548306
#> Argentina                1350610000    130670000   3570810000
#> Armenia                   142227060            .     45292403
#> Aruba                      22114096        27866    247741404
#> Australia                1851800000    817700000   7588560000
#> Azerbaijan                        .            .    319384000
#> Bahamas                           .            .    207445209
#> Bangladesh                179600000            .    481100000
#> Belarus                   967700000     27800000    520100000
#> Belize                     14346052            .     55106001
#> Bhutan                            .      1561318            .
#> Bolivia                    32734685     24183113     35485514
#> Bosnia Herzegovina         64985929     13099892      9642187
#> Botswana                   11889358       590020    110893891
#> Brazil                            .    650833690  16423399571
#> Brunei Darussalam                 .            .     15487197
#> Burundi                      509276          242      7365959
#> Cambodia                          .      1510706     80295784
#> Cameroon                  119600000       600000    276800000
#> Chile                     378561030     43437446   2525481802
#> China                   25432487660   1161195655  57956710817
#> China, Macao SAR            9277415            .    397190461
#> Colombia                  130367281     46378491    890239820
#> Congo                             .            .      6688113
#> Costa Rica               1101485483            .   2911593540
#> Dem. Rep. of the Congo            .            .      6688113
#> Djibouti                          .            .     10831645
#> Dominica                          .            .      2465954
#> Dominican Rep.             47300000            .    336100000
#> Ecuador                           .            .            .
#> El Salvador                66490425     55829522     69172931
#> Georgia                    26657177       795201     61479515
#> Greece                    580970250     78569310   1812073630
#> Grenada                     9323467         8050     14462690
#> Indonesia                 221803033     46047377   5301909818
#> Israel                  10122430000   1623010000  15177960000
#> Japan                             .  39013133686  38667607996
#> Jordan                     40620592     22143865    271086037
#> Kuwait                            .            .     34368640
#> Kyrgyzstan                  3111263      2424800     39649144
#> Lebanon                   175000000     16000000   2636000000
#> Lesotho                       20000       190000      4590000
#> Madagascar                 55482820      3121203     58285491
#> Mauritius                  70426586      2307223    592449935
#> Mexico                      7128128      6855288     24597817
#> Mongolia                    4448511      1921813    155762867
#> Morocco                   595771811      9713726   1727318889
#> Mozambique                        .            .     23173926
#> Myanmar                   219715059     11332060    403483552
#> Namibia                     2197804       317907     25172725
#> New Zealand               622455594    310878495   1323853367
#> Philippines              5183155911      9490153  15285996534
#> Poland                   4863618030    447884410  10880642069
#> Qatar                             .            .            .
#> Rep. of Korea            3569600000   6622400000  20647800000
#> Rep. of Moldova            82220000      6240000     94560000
#> Russian Federation       3063481554    608614347  12945804952
#> Samoa                        436763       105291      6555340
#> Sao Tome and Principe        639986            .      9462677
#> Saudi Arabia              272826667            .     20800000
#> Serbia                    597000000     38000000   1233000000
#> Seychelles                        .      1303764    707524869
#> Singapore               10199233148   7392297991  40090946527
#> Slovenia                  240553150     77999220   1232399684
#> Solomon Isds                  56309      1247705     12722057
#> South Africa              574897312    109260517   1939937202
#> Sri Lanka                 858350439            .     41963358
#> Suriname                     755632        36309     51543263
#> Tajikistan                        .            .       570850
#> TFYR of Macedonia         106647413      7713684    301031019
#> Thailand                  522930000     69130000  10970620000
#> Trinidad and Tobago          381884      1720187     71884458
#> Tunisia                    55004425     22774906    267337339
#> Uganda                     30298816      1527255    198691988
#> United Rep. of Tanzania    30298816      1527255    198691988
#> Uruguay                   164795392     32547155   2239791510
#> USA                     24230000000 124453000000 137854000000
#> Zambia                            .            .      4962965

services_trade_2016$services_trade_2016_tibble

#> # A tibble: 736 x 3
#>    country    product               value
#>    <chr>      <chr>                 <dbl>
#>  1 Albania    1 Transportation  200490000
#>  2 Algeria    1 Transportation  672372922
#>  3 Angola     1 Transportation   32634214
#>  4 Azerbaijan 1 Transportation 1095425000
#>  5 Argentina  1 Transportation 2175160000
#>  6 Australia  1 Transportation 4340600000
#>  7 Bahamas    1 Transportation   95553064
#>  8 Bangladesh 1 Transportation  412300000
#>  9 Armenia    1 Transportation  188458900
#> 10 Bhutan     1 Transportation 2962863685
#> # … with 726 more rows