World Trade Averages for the Period 1998-2000

A data frame that summarises all the products that different countries exported to the rest of the world. This data uses the SITC revision 2 classification with four digits product codes. The unit is year 2000 USD.

world_trade_avg_1998_to_2000

Format

A data frame with 124,336 rows and 3 columns.

Examples

world_trade_avg_1998_to_2000

#> # A tibble: 124,336 x 3
#>    country product  value
#>    <chr>   <chr>    <dbl>
#>  1 afg     0011     30068
#>  2 afg     0012     16366
#>  3 afg     0111     19273
#>  4 afg     0112       893
#>  5 afg     0113       350
#>  6 afg     0116      1561
#>  7 afg     0223       851
#>  8 afg     0224     12884
#>  9 afg     0230    114673
#> 10 afg     0240     67796
#> # … with 124,326 more rows