Complexity Measures

Complexity Measures

complexity_measures(revealed_comparative_advantage = NULL,
  c = "country", p = "product", v = "value", method = "fitness",
  iterations = 20, extremality = 1, keep_atlas = FALSE,
  tbl_output = FALSE)

Arguments

revealed_comparative_advantage
matrix or tibble/data.frame (e.g. the output of revealed_comparative_advantage()). If the input is a matrix it must be a zero/one matrix with countries in rows and products in columns. If the input is a tibble/data.frame it must contain at least three columns with countries, products and values.

c
string to indicate the column that contains exporting countries in revealed_comparative_advantage (set to "country" by default)

p
string to indicate the column that contains exported products in revealed_comparative_advantage (set to "product" by default)

v
string to indicate the column that contains traded values in revealed_comparative_advantage (set to "value" by default)

method
string to indicate to use one of these methods: reflections, eigenvalues or fitness (set to "fitness" by default)

iterations
number of iterations to use in the reflections method (set to 20 by default)

extremality
numeric coefficient to use in the fitness method (set to 1 by default)

keep_atlas
logical value to remove the countries not ranked in The Atlas of Economic Complexity (set to FALSE by default)

tbl_output
logical value to use tibble output instead of a matrix output (set to FALSE by default)

References

For more information on complexity measures, indices and its applications see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press.

Mariani M, Vidmer A, Medo M, Zhang Y (2015). “Measuring economic complexity of countries and products: which metric to use?” The European Physical Journal B, 88(11), 293. doi: 10.1140/epjb/e2015-60298-7 .

Examples

cm <- complexity_measures(package_output_demo$revealed_comparative_advantage_matrix)