Complexity Measures

complexity_measures() computes the Economic Complexity Index and the Product Complexity Index.

complexity_measures(
  balassa_index,
  method = "fitness",
  iterations = 20,
  extremality = 1
)

Arguments

balassa_index
(Type: dgCMatrix) the output from balassa_index()) or an equivalent arrangement.

method
(Type: character) one of these methods: fitness, reflections or eigenvalues. By default this is set to "fitness".

iterations
(Type: numeric) the number of iterations to use. By default this is set to 20.

extremality
(Type: numeric) the parameter to use in the fitness method. The other methods don't use this parameter. By default this is set to 1.

Value

A list of two named numeric vectors.

Details

The current implementation follows (Mariani et al. 2015) to obtain different alternatives that account for diversification in bipartite relations.

References

For more information on this index see:

Mariani M, Vidmer A, Medo M, Zhang Y (2015). “Measuring economic complexity of countries and products: which metric to use?” The European Physical Journal B, 88(11), 293. doi: 10.1140/epjb/e2015-60298-7 .

and the references therein.

Examples

complexity_measures(economiccomplexity_output$balassa_index)

#> $complexity_index_country
#>                     Bepis 9                  Bird World 
#>                0.0115128434                0.0001097279 
#>                  Crepktct 4                 Earth C-137 
#>                0.0212756540                8.3417128754 
#> Galactic Federation Outpost  Galactic Federation Prison 
#>                0.0081891360                0.0059094230 
#>                  Gear World                Glaagablaaga 
#>                0.1670017470                0.4440184218 
#>              Gromflom Prime 
#>                0.0002701715 
#> 
#> $complexity_index_product
#>        alpha         beta        delta      epsilon          eta        gamma 
#> 5.7151134328 0.0003691566 5.7151134328 0.0039046311 0.0935850244 0.0001104686 
#>         iota        kappa       lambda           mu        theta         zeta 
#> 0.3405363984 0.0049668423 0.0181804513 0.0920772586 0.0088859916 0.0071569113 
#>