Complexity Measures

Complexity Measures

complexity_measures(rca = NULL, country = "country",
  product = "product", value = "value", method = "fitness",
  iterations = 20, extremality = 1, keep_atlas = FALSE,
  tbl_output = FALSE)

Arguments

rca
matrix or tibble/data.frame (e.g. the output of revealed_comparative_advantage()). If the input is a matrix it must be a zero/one matrix with countries in rows and products in columns. If the input is a tibble/data.frame it must contain at least three columns with countries, products and values.

country
string to indicate the column that contains exporting countries in rca (set to "country" by default)

product
string to indicate the column that contains exported products in rca (set to "product" by default)

value
string to indicate the column that contains traded values in rca (set to "value" by default)

method
string to indicate to use one of these methods: reflections, eigenvalues or fitness (set to "fitness" by default)

iterations
number of iterations to use in the reflections method (set to 20 by default)

extremality
numeric coefficient to use in the fitness method (set to 1 by default)

keep_atlas
logical value to remove the countries not ranked in The Atlas of Economic Complexity (set to FALSE by default)

tbl_output
logical value to use tibble output instead of a matrix output (set to FALSE by default)

References

For more information on complexity measures, indices and its applications see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

Mariani M, Vidmer A, Medo M, Zhang Y (2015). “Measuring economic complexity of countries and products: which metric to use?” The European Physical Journal B, 88(11), 293. doi: 10.1140/epjb/e2015-60298-7 .

Examples

complexity_measures(
 rca = package_output_demo$revealed_comparative_advantage_matrix,
 tbl_output = TRUE
)

#> $economic_complexity_index
#> # A tibble: 80 x 2
#>    country                value
#>    <chr>                  <dbl>
#>  1 USA                39.6     
#>  2 Japan              39.2     
#>  3 Singapore           0.422   
#>  4 Lebanon             0.422   
#>  5 South Africa        0.422   
#>  6 China               0.000167
#>  7 Poland              0.000156
#>  8 Russian Federation  0.000156
#>  9 Serbia              0.000156
#> 10 TFYR of Macedonia   0.000156
#> # … with 70 more rows
#> 
#> $product_complexity_index
#> # A tibble: 11 x 2
#>    product                                                  value
#>    <chr>                                                    <dbl>
#>  1 8 Royalties and license fees                     10.9         
#>  2 6 Financial services                              0.0835      
#>  3 7 Computer and information services               0.00000450  
#>  4 9 Other business services                         0.00000401  
#>  5 5 Insurance services                              0.00000263  
#>  6 4 Construction services                           0.00000238  
#>  7 10 Personal, cultural, and recreational services  0.00000116  
#>  8 3 Communications services                         0.000000499 
#>  9 1 Transportation                                  0.000000488 
#> 10 11 Government services, n.i.e.                    0.000000417 
#> 11 2 Travel                                          0.0000000716
#> 
#> $diversity
#> # A tibble: 80 x 2
#>    country    value
#>    <chr>      <dbl>
#>  1 Albania        3
#>  2 Algeria        5
#>  3 Angola         1
#>  4 Argentina      5
#>  5 Armenia        6
#>  6 Aruba          1
#>  7 Australia      3
#>  8 Azerbaijan     2
#>  9 Bahamas        1
#> 10 Bangladesh     4
#> # … with 70 more rows
#> 
#> $ubiquity
#> # A tibble: 11 x 2
#>    product                                          value
#>    <chr>                                            <dbl>
#>  1 1 Transportation                                    43
#>  2 10 Personal, cultural, and recreational services    29
#>  3 11 Government services, n.i.e.                      38
#>  4 2 Travel                                            53
#>  5 3 Communications services                           41
#>  6 4 Construction services                             21
#>  7 5 Insurance services                                18
#>  8 6 Financial services                                 4
#>  9 7 Computer and information services                 17
#> 10 8 Royalties and license fees                         2
#> 11 9 Other business services                           19
#>