Complexity Outlook

complexity_outlook() computes the Complexity Outlook Index and the Complexity Outlook Gain.

complexity_outlook(balassa_index, proximity_product, complexity_index_product)

Arguments

balassa_index
(Type: dgCMatrix) the output from balassa_index()) or an equivalent arrangement.

proximity_product
(Type: dgCMatrix) the output from proximity()) or an equivalent arrangement.

complexity_index_product
(Type: numeric) the output from complexity_measures()) or an equivalent arrangement.

Value

A list of two named numeric vectors.

Details

The current implementation follows (Hausmann et al. 2014) to measure unexploited export oppportunities.

References

For more information on this index see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 , https://doi.org/10.7551/mitpress/9647.001.0001.

and the references therein.

Examples

co <- complexity_outlook(
 economiccomplexity_output$balassa_index,
 economiccomplexity_output$proximity$proximity_product,
 economiccomplexity_output$complexity_measures$complexity_index_product
)

# partial view of complexity outlook
co$complexity_outlook_index[1:5]

#>        afg        ago        aia        alb        and 
#> 103.948610   9.962401 122.311158 152.107317 151.295380 

co$complexity_outlook_gain[1:5,1:5]

#> 5 x 5 Matrix of class "dgeMatrix"
#>          0011      0012      0013      0014      0015
#> afg 0.8615531 0.7613878 0.7537907 1.0961458 0.8143851
#> ago 0.9681802 0.8436855 0.8219918 1.2151040 0.8883146
#> aia 0.8339114 0.7475080 0.7247198 1.0425043 0.7671877
#> alb 0.7979779 0.7199705 0.7132042 1.0170093 0.7636545
#> and 0.0000000 0.7118512 0.6829531 0.9987236 0.0000000