Example outputs of the functions within the package

List of different economic complexity computations obtained by using the functions within the package. The output is provided as a list of data frames.

ec_output_demo

Format

A list of data frames with 3 elements.

rca

`rca()` output (single data frame)

complexity

`complexity()` output (list of four data frames)

value

`proximity()` output (list of two data frames)

Examples

ec_output_demo

#> $rca
#> # A tibble: 44,435 x 3
#>    country product value
#>    <chr>   <chr>   <dbl>
#>  1 afg     0106        0
#>  2 afg     0405        0
#>  3 afg     0407        0
#>  4 afg     0501        1
#>  5 afg     0504        1
#>  6 afg     0703        0
#>  7 afg     0712        0
#>  8 afg     0714        1
#>  9 afg     0801        1
#> 10 afg     0802        1
#> # … with 44,425 more rows
#> 
#> $complexity
#> $complexity$complexity_c
#> # A tibble: 99 x 2
#>    country value
#>    <chr>   <dbl>
#>  1 deu     19.5 
#>  2 fra     12.0 
#>  3 usa     11.6 
#>  4 gbr      8.43
#>  5 che      6.92
#>  6 nld      6.63
#>  7 ita      6.55
#>  8 swe      4.85
#>  9 bel      4.32
#> 10 aut      4.29
#> # … with 89 more rows
#> 
#> $complexity$complexity_p
#> # A tibble: 991 x 2
#>    product value
#>    <chr>   <dbl>
#>  1 2702     48.1
#>  2 7111     48.1
#>  3 2850     29.7
#>  4 8804     18.4
#>  5 2811     17.9
#>  6 9307     17.9
#>  7 9016     12.7
#>  8 9112     12.7
#>  9 8701     12.1
#> 10 8707     12.1
#> # … with 981 more rows
#> 
#> $complexity$diversity
#> # A tibble: 99 x 2
#>    country value
#>    <chr>   <dbl>
#>  1 ago        52
#>  2 alb        32
#>  3 are         1
#>  4 arg        83
#>  5 aus        81
#>  6 aut       290
#>  7 bel       320
#>  8 bgr       127
#>  9 bol        27
#> 10 bra        70
#> # … with 89 more rows
#> 
#> $complexity$ubiquity
#> # A tibble: 991 x 2
#>    product value
#>    <chr>   <dbl>
#>  1 0101       13
#>  2 0102       19
#>  3 0103       13
#>  4 0104       12
#>  5 0105        9
#>  6 0106       27
#>  7 0201       17
#>  8 0203       12
#>  9 0204        4
#> 10 0205       10
#> # … with 981 more rows
#> 
#> 
#> $proximity
#> $proximity$proximity_c
#> # A tibble: 4,484 x 3
#>    from  to     value
#>    <chr> <chr>  <dbl>
#>  1 alb   ago   0.0769
#>  2 are   ago   0.0192
#>  3 arg   ago   0.133 
#>  4 aus   ago   0.0741
#>  5 aut   ago   0.0241
#>  6 bel   ago   0.0281
#>  7 bgr   ago   0.110 
#>  8 bol   ago   0.0385
#>  9 bra   ago   0.2   
#> 10 can   ago   0.0845
#> # … with 4,474 more rows
#> 
#> $proximity$proximity_p
#> # A tibble: 408,869 x 3
#>    from  to    value
#>    <chr> <chr> <dbl>
#>  1 0102  0101  0.316
#>  2 0103  0101  0.308
#>  3 0104  0101  0.385
#>  4 0106  0101  0.148
#>  5 0201  0101  0.353
#>  6 0203  0101  0.462
#>  7 0204  0101  0.154
#>  8 0205  0101  0.385
#>  9 0206  0101  0.385
#> 10 0207  0101  0.429
#> # … with 408,859 more rows
#> 
#>