Productivity Levels

Productivity Levels

productivity_levels(trade_data = NULL, country1 = "country",
  product1 = "product", value1 = "value", gdp_data = NULL,
  country2 = "country", value2 = "value", tbl_output = FALSE)

Arguments

trade_data
matrix or tibble/data.frame (e.g. world_trade_2017). If the input is a matrix it must be a zero/one matrix with countries in rows and products in columns. If the input is a tibble/data.frame it must contain at least three columns with countries, products and values.

country1
string to indicate the column that contains exporting countries in revealed_comparative_advantage (set to "country" by default)

product1
string to indicate the column that contains exported products in revealed_comparative_advantage (set to "product" by default)

value1
string to indicate the column that contains traded values in revealed_comparative_advantage (set to "value" by default)

gdp_data
vector or tibble/data.frame (e.g. world_gdp_and_population_2017). If the input is a vector it must be a numeric vector with optional names. If the input is a tibble/data.frame it must contain at least two columns with countries and values.

country2
string to indicate the column that contains exporting countries in revealed_comparative_advantage (set to "country" by default)

value2
string to indicate the column that contains traded values in revealed_comparative_advantage (set to "value" by default)

tbl_output
when set to TRUE the output will be a tibble instead of a matrix (default set to FALSE)

References

For more information on prody and its applications see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

Hausmann R, Hwang J, Rodrik D (2005). “What You Export Matters.” Working Paper 11905, National Bureau of Economic Research. doi: 10.3386/w11905 .

Examples

productivity_levels(
 trade_data = services_trade_2016$services_trade_2016_matrix,
 country1 = "country",
 product1 = "product",
 value1 = "value",
 gdp_data = gdp_pc_2016$gdp_pc_2016_numeric,
 country2 = "country",
 value2 = "value",
 tbl_output = TRUE
)

#> $economies_productivity_level
#> # A tibble: 80 x 2
#>    name        value
#>    <chr>       <dbl>
#>  1 Albania    27854.
#>  2 Algeria    23580.
#>  3 Angola     11020.
#>  4 Argentina  11714.
#>  5 Armenia     7948.
#>  6 Aruba      10359.
#>  7 Australia  11903.
#>  8 Azerbaijan 12047.
#>  9 Bahamas    13998.
#> 10 Bangladesh 15313.
#> # … with 70 more rows
#> 
#> $products_productivity_level
#> # A tibble: 11 x 2
#>    name                                              value
#>    <chr>                                             <dbl>
#>  1 1 Transportation                                 11104.
#>  2 10 Personal, cultural, and recreational services 12311.
#>  3 11 Government services, n.i.e.                    4902.
#>  4 2 Travel                                         11058.
#>  5 3 Communications services                        11285.
#>  6 4 Construction services                           8861.
#>  7 5 Insurance services                             14503.
#>  8 6 Financial services                             20371.
#>  9 7 Computer and information services              12564.
#> 10 8 Royalties and license fees                     33531.
#> 11 9 Other business services                        12736.
#>