Projections of a Country-Product Network

proximity() computes two graphs that are particularly useful to visualize product-product and country-country similarity.

projections(
  proximity_country,
  proximity_product,
  avg_links = 5,
  tolerance = 0.05,
  compute = "both"
)

Arguments

proximity_country
(Type: dgCMatrix) the output from proximity()) or an equivalent arrangement.

proximity_product
(Type: dgCMatrix) the output from proximity()) or an equivalent arrangement.

avg_links
average number of connections for the projections. By default this is set to 5.

tolerance
tolerance for proximity variation on each iteration until obtaining the desired average number of connections. By default this is set to 0.05.

compute
(Type: character) the proximity to compute. By default this is "both" (both projections) but it can also be "country" or "product".

Value

A list of two graphs.

Details

The current implementation follows (Hausmann et al. 2014) to create simplified graphs that correspond to a simplification of the proximity matrices. The result is obtained by iterating and reducing links until the desired average number of links per node is obtained, or a spaning tree after the strongest links is returned when is not possible to return the desired network.

References

For more information see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

and the references therein.

Examples

projections(
  proximity_country = economiccomplexity_output$proximity$proximity_country,
  proximity_product = economiccomplexity_output$proximity$proximity_product
)

#> computing product projection...

#> --------------------------------------------------

#> 0 threshold...

#> 0.05 threshold achieves the avg number of connections

#> computing product projection...

#> --------------------------------------------------

#> 0 threshold...

#> 0.05 threshold...

#> 0.1 threshold...

#> 0.15 threshold...

#> 0.2 threshold...

#> 0.25 threshold...

#> 0.3 threshold achieves the avg number of connections

#> $network_country
#> IGRAPH 372fbcc UNW- 9 21 -- 
#> + attr: name (v/c), weight (e/n)
#> + edges from 372fbcc (vertex names):
#>  [1] Gear World                 --Glaagablaaga               
#>  [2] Galactic Federation Prison --Gromflom Prime             
#>  [3] Galactic Federation Prison --Glaagablaaga               
#>  [4] Galactic Federation Outpost--Glaagablaaga               
#>  [5] Galactic Federation Outpost--Gear World                 
#>  [6] Galactic Federation Outpost--Galactic Federation Prison 
#>  [7] Earth C-137                --Glaagablaaga               
#>  [8] Earth C-137                --Gear World                 
#> + ... omitted several edges
#> 
#> $network_product
#> IGRAPH 3732c7e UNW- 12 27 -- 
#> + attr: name (v/c), weight (e/n)
#> + edges from 3732c7e (vertex names):
#>  [1] theta  --zeta   mu     --zeta   mu     --theta  lambda --zeta  
#>  [5] lambda --theta  lambda --mu     kappa  --mu     kappa  --lambda
#>  [9] iota   --mu     iota   --kappa  gamma  --kappa  eta    --theta 
#> [13] eta    --mu     eta    --kappa  eta    --iota   epsilon--theta 
#> [17] epsilon--mu     delta  --zeta   delta  --theta  delta  --mu    
#> [21] delta  --lambda beta   --zeta   alpha  --zeta   alpha  --theta 
#> [25] alpha  --mu     alpha  --lambda alpha  --delta 
#>