Country-product bipartite network projections

rca computes complexity indices following the definitions from Mariani et al. (2015)

projections(proximity_c, proximity_p, cutoff_c = 0.2, cutoff_p = 0.4,
  compute = "both", tbl = TRUE, from_c = "from", to_c = "to",
  value_c = "value", from_p = "from", to_p = "to",
  value_p = "value")

Arguments

proximity_c
matrix or data frame with product proximity values

proximity_p
matrix or data frame with country proximity values

cutoff_c
all the links with a proximity below this value will be removed from the country projection (by default is 0.2)

cutoff_p
all the links with a proximity below this value will be removed from the product projection (by default is 0.4)

compute
"country", "product" or "both" (default) projections

tbl
TRUE (default) returns a data.frame and FALSE returns a matrix

from_c
column containing origin (applies only if proximity_c is a data.frame)

to_c
column containing destination (applies only if proximity_c is a data.frame)

value_c
column containing proximity values (applies only if proximity_c is a data.frame)

from_p
column containing origin (applies only if proximity_p is a data.frame)

to_p
column containing destination (applies only if proximity_p is a data.frame)

value_p
column containing proximity values (applies only if proximity_p is a data.frame)

Value

A list with two data frames or matrices.

Details

Given a matrix of \(C\times C\) with country proximity values and a matrix of \(P\times P\) with product proximity values, this function obtains a simplified network in two steps: 1) creates a network skeleton by appliying the minimum spanning tree algorithm from the igraph package, but multiplying proximities by minus one, so that the strongest links are considered for the network skeleton 2) append additional links to the skeleton by aading the links with proximity values above a user-defined cutoff

References

For more information about visualizing bipartite networks projections see:

Goh K, Cusick ME, Valle D, Childs B, Vidal M, Barabási A (2007). “The human disease network.” Proceedings of the National Academy of Sciences, 104(21), 8685--8690. ISSN 0027-8424, doi: 10.1073/pnas.0701361104 , https://www.pnas.org/content/104/21/8685.full.pdf, https://www.pnas.org/content/104/21/8685.

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

and the references therein.

See also

Examples

projections(
  ec_output_demo$proximity$proximity_c,
  ec_output_demo$proximity$proximity_p
)

#> $projection_c
#> # A tibble: 497 x 3
#>    from  to    value
#>    <chr> <chr> <dbl>
#>  1 khm   cmr   0.2  
#>  2 khm   ecu   0.2  
#>  3 khm   gtm   0.225
#>  4 khm   idn   0.214
#>  5 khm   mdg   0.204
#>  6 khm   mli   0.325
#>  7 khm   nic   0.2  
#>  8 khm   pak   0.241
#>  9 khm   sdn   0.2  
#> 10 khm   nga   0.267
#> # … with 487 more rows
#> 
#> $projection_p
#> # A tibble: 56,160 x 3
#>    from  to    value
#>    <chr> <chr> <dbl>
#>  1 0711  9614  0.467
#>  2 0711  0103  0.4  
#>  3 0711  0203  0.467
#>  4 0711  0207  0.4  
#>  5 0711  0210  0.4  
#>  6 0711  0405  0.4  
#>  7 0711  0407  0.5  
#>  8 0711  0505  0.467
#>  9 0711  0701  0.529
#> 10 0711  0702  0.467
#> # … with 56,150 more rows
#>