Proximity

proximity computes proximity

proximity(rca, diversity, ubiquity, country_r = "country",
  product_r = "product", value_r = "value", country_d = "country",
  value_d = "value", product_u = "product", value_u = "value",
  tbl = TRUE, compute = "both")

Arguments

rca
matrix or data.frame with RCA values

diversity
matrix or data.frame with diversity values

ubiquity
matrix or data.frame with ubiquity values

country_r
column containing countries (applies only if d is a data.frame)

product_r
column containing products (applies only if d is a data.frame)

value_r
column containing traded values (applies only if d is a data.frame)

country_d
column containing countries (applies only if d is a data.frame)

value_d
column containing values (applies only if d is a data.frame)

product_u
column containing products (applies only if d is a data.frame)

value_u
column containing values (applies only if d is a data.frame)

tbl
TRUE (default) returns a data.frame and FALSE returns a matrix

compute
"country", "product" or "both" (default) matrices

Value

A list with two data frames or matrices.

References

For more information on proximity and its applications see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

and the references therein.

See also

Examples

proximity(
  rca = ec_output_demo$rca_tbl,
  d = ec_output_demo$complexity_measures_tbl$diversity,
  u = ec_output_demo$complexity_measures_tbl$ubiquity
)

#> $product_proximity
#> # A tibble: 4,484 x 3
#>    from  to     value
#>    <chr> <chr>  <dbl>
#>  1 alb   ago   0.0769
#>  2 are   ago   0.0192
#>  3 arg   ago   0.133 
#>  4 aus   ago   0.0741
#>  5 aut   ago   0.0241
#>  6 bel   ago   0.0281
#>  7 bgr   ago   0.110 
#>  8 bol   ago   0.0385
#>  9 bra   ago   0.2   
#> 10 can   ago   0.0845
#> # ... with 4,474 more rows
#> 
#> $country_proximity
#> # A tibble: 408,869 x 3
#>    from  to    value
#>    <chr> <chr> <dbl>
#>  1 0102  0101  0.316
#>  2 0103  0101  0.308
#>  3 0104  0101  0.385
#>  4 0106  0101  0.148
#>  5 0201  0101  0.353
#>  6 0203  0101  0.462
#>  7 0204  0101  0.154
#>  8 0205  0101  0.385
#>  9 0206  0101  0.385
#> 10 0207  0101  0.429
#> # ... with 408,859 more rows
#>