Proximity

Proximity

proximity(rca = NULL, country = "country", product = "product",
  value = "value", diversity = NULL, diversity_country = "country",
  diversity_value = "value", ubiquity = NULL,
  ubiquity_product = "product", ubiquity_value = "value",
  tbl_output = FALSE, compute = "both")

Arguments

rca
matrix or tibble/data.frame in long format (e.g. the output of revealed_comparative_advantage()). If it is a matrix it must be a zero/one matrix with countries in the row names and products in the column names. If rca is a tibble/data.frame it must contain the columns country (character/ factor), product (character/factor) and discrete RCA (integer)

country
string to indicate the column that contains exporting countries (default set to "country" that is the output of revealed_comparative_advantage(), applies only if rca is a data.frame)

product
string to indicate the column that contains exported products (default set to "product" that is the output of revealed_comparative_advantage(), applies only if rca is a data.frame)

value
string to indicate the column that contains RCA values (default set to "value" that is the output of revealed_comparative_advantage(), applies only if rca is a data.frame)

diversity
numeric vector or tibble/data.frame containing diversity measures (e.g. diversity from complexity_measures())

diversity_country
string to indicate the column that contains diversity countries (default set to "country" that is the output of economic_complexity_measures())

diversity_value
string to indicate the column that contains diversity values (default set to "value" that is the output of economic_complexity_measures())

ubiquity
numeric vector or tibble/data.frame containing ubiquity measures (e.g. ubiquity from complexity_measures())

ubiquity_product
string to indicate the column that contains diversity products (default set to "product" that is the output of economic_complexity_measures())

ubiquity_value
string to indicate the column that contains ubiquity values (default set to "value" that is the output of economic_complexity_measures())

tbl_output
when set to TRUE the output will be a tibble instead of a matrix (default set to FALSE)

compute
by default set to "both", it can also be "country" or "product"

References

For more information on proximity and its applications see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

Examples

proximity(
 rca = package_output_demo$revealed_comparative_advantage_matrix,
 diversity = package_output_demo$complexity_measures_numeric$diversity,
 ubiquity = package_output_demo$complexity_measures_numeric$ubiquity,
 tbl_output = TRUE
)

#> $proximity_countries
#> # A tibble: 2,683 x 3
#>    from       to      value
#>    <chr>      <chr>   <dbl>
#>  1 Angola     Albania 0.333
#>  2 Argentina  Albania 0.4  
#>  3 Armenia    Albania 0.5  
#>  4 Aruba      Albania 0.333
#>  5 Australia  Albania 1    
#>  6 Azerbaijan Albania 0.333
#>  7 Bahamas    Albania 0.333
#>  8 Bangladesh Albania 0.25 
#>  9 Belarus    Albania 0.25 
#> 10 Belize     Albania 0.667
#> # … with 2,673 more rows
#> 
#> $proximity_products
#> # A tibble: 51 x 3
#>    from                                             to                value
#>    <chr>                                            <chr>             <dbl>
#>  1 10 Personal, cultural, and recreational services 1 Transportation 0.326 
#>  2 11 Government services, n.i.e.                   1 Transportation 0.512 
#>  3 2 Travel                                         1 Transportation 0.434 
#>  4 3 Communications services                        1 Transportation 0.581 
#>  5 4 Construction services                          1 Transportation 0.349 
#>  6 5 Insurance services                             1 Transportation 0.279 
#>  7 6 Financial services                             1 Transportation 0.0233
#>  8 7 Computer and information services              1 Transportation 0.256 
#>  9 8 Royalties and license fees                     1 Transportation 0.0233
#> 10 9 Other business services                        1 Transportation 0.279 
#> # … with 41 more rows
#>