Proximity

proximity() computes two matrices that account for the similarity between pairs of countries and pairs of products.

proximity(balassa_index, compute = "both")

Arguments

balassa_index
(Type: dgCMatrix) the output from balassa_index()) or an equivalent arrangement.

compute
(Type: character) the proximity to compute. By default this is "both" (both proximity matrices) but it can also be "country" or "product".

Value

A list of two matrices.

Details

The current implementation follows (Hausmann et al. 2014) to obtain the likelihood that two products "p1" and "p2" are exported by the same country and, conversely, that two countries "c1" and "c2" export the same product.

References

For more information see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

and the references therein.

Examples

proximity(economiccomplexity_output$balassa_index)

#> $proximity_country
#> 9 x 9 sparse Matrix of class "dsCMatrix"
#>                               Bepis 9 Bird World Crepktct 4 Earth C-137
#> Bepis 9                     1.0000000  .          .           0.2857143
#> Bird World                  .          1.0000000  0.3333333   .        
#> Crepktct 4                  .          0.3333333  1.0000000   0.1428571
#> Earth C-137                 0.2857143  .          0.1428571   1.0000000
#> Galactic Federation Outpost 0.3333333  0.3333333  0.6666667   0.1428571
#> Galactic Federation Prison  .          0.3333333  0.3333333   0.1428571
#> Gear World                  0.5000000  .          .           0.4285714
#> Glaagablaaga                .          0.2000000  0.4000000   0.1428571
#> Gromflom Prime              .          .          .           .        
#>                             Galactic Federation Outpost
#> Bepis 9                                       0.3333333
#> Bird World                                    0.3333333
#> Crepktct 4                                    0.6666667
#> Earth C-137                                   0.1428571
#> Galactic Federation Outpost                   1.0000000
#> Galactic Federation Prison                    0.3333333
#> Gear World                                    0.2500000
#> Glaagablaaga                                  0.4000000
#> Gromflom Prime                                .        
#>                             Galactic Federation Prison Gear World Glaagablaaga
#> Bepis 9                                      .          0.5000000    .        
#> Bird World                                   0.3333333  .            0.2000000
#> Crepktct 4                                   0.3333333  .            0.4000000
#> Earth C-137                                  0.1428571  0.4285714    0.1428571
#> Galactic Federation Outpost                  0.3333333  0.2500000    0.4000000
#> Galactic Federation Prison                   1.0000000  .            0.2000000
#> Gear World                                   .          1.0000000    0.4000000
#> Glaagablaaga                                 0.2000000  0.4000000    1.0000000
#> Gromflom Prime                               0.3333333  .            .        
#>                             Gromflom Prime
#> Bepis 9                          .        
#> Bird World                       .        
#> Crepktct 4                       .        
#> Earth C-137                      .        
#> Galactic Federation Outpost      .        
#> Galactic Federation Prison       0.3333333
#> Gear World                       .        
#> Glaagablaaga                     .        
#> Gromflom Prime                   1.0000000
#> 
#> $proximity_product
#> 12 x 12 sparse Matrix of class "dsCMatrix"

#>    [[ suppressing 12 column names ‘alpha’, ‘beta’, ‘delta’ ... ]]

#>                                                                       
#> alpha   1.0000000 .   1.0000000 0.25 .         .   .         .        
#> beta    .         1.0 .         .    .         0.2 .         .        
#> delta   1.0000000 .   1.0000000 0.25 .         .   .         .        
#> epsilon 0.2500000 .   0.2500000 1.00 0.2500000 0.2 .         0.2500000
#> eta     .         .   .         0.25 1.0000000 0.2 0.5000000 0.3333333
#> gamma   .         0.2 .         0.20 0.2000000 1.0 0.2000000 0.6000000
#> iota    .         .   .         .    0.5000000 0.2 1.0000000 0.3333333
#> kappa   .         .   .         0.25 0.3333333 0.6 0.3333333 1.0000000
#> lambda  0.5000000 .   0.5000000 0.25 .         0.2 .         0.3333333
#> mu      0.3333333 .   0.3333333 0.50 0.6666667 0.2 0.3333333 0.3333333
#> theta   0.3333333 .   0.3333333 0.75 0.3333333 .   .         .        
#> zeta    0.5000000 0.5 0.5000000 0.25 .         0.2 .         .        
#>                                                
#> alpha   0.5000000 0.3333333 0.3333333 0.5000000
#> beta    .         .         .         0.5000000
#> delta   0.5000000 0.3333333 0.3333333 0.5000000
#> epsilon 0.2500000 0.5000000 0.7500000 0.2500000
#> eta     .         0.6666667 0.3333333 .        
#> gamma   0.2000000 0.2000000 .         0.2000000
#> iota    .         0.3333333 .         .        
#> kappa   0.3333333 0.3333333 .         .        
#> lambda  1.0000000 0.3333333 0.3333333 0.5000000
#> mu      0.3333333 1.0000000 0.6666667 0.3333333
#> theta   0.3333333 0.6666667 1.0000000 0.3333333
#> zeta    0.5000000 0.3333333 0.3333333 1.0000000
#>